feat: Add validation in signup user

+++ backend/controllers/userControllers.js
@@ -5,17 +5,12 @@ function generateUUID(req, res) {
         res.status(200).json(uuid)
 }
 
+const { default: mongoose } = require('mongoose')
 const User = require('../models/userModel')
 
 async function signupUser(req, res) {
         try {
                 const { email, password } = req.body
-
-                if (!email || !password) {
-                        res.status(400).json({error: 'Missing required fields'})
-                        return
-                }
-
                 const user = await User.signup(email, password)
                 res.status(200).json(user)
         }
@@ -31,9 +26,19 @@ async function loginUser(req, res) {
 
 async function deleteUser(req, res) {
         const { id } = req.params
+        console.log(id)
 
-        // check if user does not exists
-        
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+                res.status(400).json({error: 'Invalid id'})
+                return
+        }
+
+        const userExists = await User.findById(id)
+        if (!userExists) {
+                res.status(404).json({error: 'User does not exists'})
+                return
+        }
+
         const user = await User.findByIdAndDelete(id)
 
         res.status(200).json(user)
